src/webscraper/glynn/requirements.txt
# Scraper runs under Python 3.6.3

tabula-py >= 1.0.0
numpy >= 1.13.1
pandas >= 0.20.3
requests >= 2.18.4